+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [5.0.0] - 2026-01-26
+
+- Change bundling process to use `@macrostrat/web-components-bundler`
+- Update peer dependencies to React 18
+- Require styles to be imported separately as
+  `@macrostrat/ui-components/styles.css` or
+  `@macrostrat/ui-components/dist/ui-components.css`
+-
+
 ## [4.6.0] - 2026-01-20
 
 - Modernized color picker component
@@ -122,7 +131,7 @@ Fixed a bug in UI box model calculations
 
 ## [4.1.2] - 2025-02-16
 
-Add body classes `light-mode` and `bp5-light` to `DarkModeProvider` when it is
+Add body classes `light-mode` and `bp6-light` to `DarkModeProvider` when it is
 explicitly disabled.
 
 ## [4.1.1] - 2025-02-15